chore: move map store helpers closer to call site
These helper methods are only used within the map.store and
should be moved closer to make the code structure easier to reason with.
