+++ src/stores/Maps/maps.store.ts
@@ -12,7 +12,7 @@ import type { Subscription } from 'rxjs'
 import { ModuleStore } from '../common/module.store'
 import { getUserAvatar } from '../User/user.store'
 import { MAP_GROUPINGS } from './maps.groupings'
-import { generatePins, generatePinDetails } from 'src/mocks/maps.mock'
+import { generatePins, generatePinDetails } from 'src/stores/Maps/generatePins'
 import type { IUserPP } from 'src/models/user_pp.models'
 import type { IUploadedFileMeta } from '../storage'
 import {
